# docs/merchant/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/merchant/api/v1/auth/login": {
            "post": {
                "description": "Аутентифицирует пользователя и возвращает токен доступа",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Вход пользователя",
                "parameters": [
                    {
                        "description": "Данные для входа",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_1stpay_1stpay_internal_transport_rest_merchant_restdto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Возвращает токен доступа",
                        "schema": {
                            "$ref": "#/definitions/github_com_1stpay_1stpay_internal_transport_rest_merchant_restdto.AccessTokenResponse"
                        }
                    },
                    "400": {
                        "description": "Неверные входные данные",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещён",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/merchant/api/v1/auth/register": {
            "post": {
                "description": "Registers new user and returns access token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User registration",
                "parameters": [
                    {
                        "description": "Registration data",
                        "name": "register",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_1stpay_1stpay_internal_transport_rest_merchant_restdto.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Returns success token",
                        "schema": {
                            "$ref": "#/definitions/github_com_1stpay_1stpay_internal_transport_rest_merchant_restdto.AccessTokenResponse"
                        }
                    },
                    "400": {
                        "description": "Incorrect data",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/merchant/api/v1/blockchain/list/": {
            "get": {
                "description": "Retrieve a list of all active blockchains.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blockchain"
                ],
                "summary": "List active blockchains",
                "responses": {
                    "200": {
                        "description": "List of active blockchains",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_1stpay_1stpay_internal_transport_rest_merchant_restdto.BlockchainListResponseDTO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gin.H": {
            "type": "object",
            "additionalProperties": {}
        },
        "github_com_1stpay_1stpay_internal_domain_enum.NetworkType": {
            "type": "string",
            "enum": [
                "evm",
                "tron",
                "ton",
                "solana"
            ],
            "x-enum-varnames": [
                "EVM",
                "TRON",
                "TON",
                "SOLANA"
            ]
        },
...
        "github_com_1stpay_1stpay_internal_transport_rest_merchant_restdto.AccessTokenResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                }
            }
        },
        "github_com_1stpay_1stpay_internal_transport_rest_merchant_restdto.BlockchainListResponseDTO": {
            "type": "object",
            "properties": {
                "chain_type": {
                    "$ref": "#/definitions/github_com_1stpay_1stpay_internal_domain_enum.NetworkType"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "logo": {
                    "$ref": "#/definitions/null.String"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "github_com_1stpay_1stpay_internal_transport_rest_merchant_restdto.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "github_com_1stpay_1stpay_internal_transport_rest_merchant_restdto.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "null.String": {
            "type": "object",
            "properties": {
                "string": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if String is not NULL",
                    "type": "boolean"
                }
            }
        }
    }
}